+++ docs/source/app.rst
@@ -1,0 +1,7 @@
+app module
+==========
+
+.. automodule:: app
+   :members:
+   :show-inheritance:
+   :undoc-members:
