+++ docs/source/index.rst
@@ -1,0 +1,18 @@
+.. Dev Log documentation master file, created by
+   sphinx-quickstart on Fri Apr 10 10:25:26 2026.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Dev Log documentation
+=====================
+
+Add your content using ``reStructuredText`` syntax. See the
+`reStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_
+documentation for details.
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   modules
