+++ docs/source/modules.rst
@@ -1,0 +1,7 @@
+app
+===
+
+.. toctree::
+   :maxdepth: 4
+
+   app
